Seperate functions by files
함수들을 파일별로 나누었습니다.
