+++ src/Button.jsx
@@ -1,0 +1,14 @@
+import React from 'react';
+
+function Button({ children, onClick }) {
+  return (
+    <button
+      type="button"
+      onClick={() => onClick({ increased: children })}
+    >
+      {children}
+    </button>
+  );
+}
+
+export default Button;
